+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Use "debug" signing config instead of creating a new one
 - Remove unnecessary lines from .gitignore
+- Remove Compose previews
 
 ## [1.8.2](https://github.com/ldeso/blitz/releases/tag/v1.8.2) – 2024-04-19